requirements.txt
flask==3.0.0
paho-mqtt==1.6.1
flask-socketio
python-socketio 